Attempt at a weird experimental pubsub system for this game
Maybe we need this later but we can try to get by without it for now
We need a model for data passing. IoC could work, or this pub sub
with an additional global scope might work. For now skip it
and focus on doing the small iterations to get a simple working
version of the game