Try out having the game own the renderer
The game's data is private. The renderer wants to show that data. Do we allow the game to return the data. Read only for the state of the game?
What if the game owned the renderer and since the game ownded the renderer the data was in scope their?
I'm not sure how the control should go but i'm messing around with some stuff here.